Redesign of StyleSet
The class StyleSet has been redesigned. It now longer returns null if
the styles are exhausted, but instead cycles back to the first style.
I also now sports hash maps to allow for storing styles for specific
names.

More work on the "FunctionJob" class has been done. The goal is to
provide a very easy-to-use base class for jobs which plot a function
obtained from the data.