New Interface: ISemanticComponent
The new interface ISemanticComponent has been added to the Experiment
Data API specification. This interface allows to build components which
have a certain meaning or name that can be written to a document as well
as which can make "suggestions" about path components for other files
created based on them.

The original idea was to have attributes (such as functions) which have
convenient methods to describe them in text. However, the interface also
lends itself to be included into all named elements of the Experiment
Data API. This way, we already have a default name of writing experiment
names or dimension names.