Improved Version of ECDF Module/FunctionJob + Several Small Improvements
The FunctionJob and the ECDF module using it have been improved. Many of
the text printed is based on the ISemanticComponent interface, which now
can also understand an instance of ETextCase provided to its appendName
method. The goal is to distinguish text used within a sentence and in a
headline.

A bug in the way the "alt" attribute in figures for XHTML are generated
has been fixed: Tags within the attribute are now automatically deleted.

The useless class AxisChart has been deleted.

Some settings used when painting charts have been modified in order to
get better-looking charts.