ETextCase Now Respects English Grammar in Title Case
The appendWord(s) methods of ETextCase now properly respect the English
grammar for title cases: articles, conjunctions, and prepositions not
occurring at the start of the title should not begin with an upper case
letter. Instead, they remain lower case.

I therefore added a lookup table with all English articles,
conjunctions, and prepositions I could find / think of. The table is
sorted and if the text case is IN_TITLE, a lookup is performed via
binary search.

Source / more info:
http://www.grammar-monster.com/lessons/capital_letters_title_case.htm