More Shared Shadow Attribute Computation
I increased the checks for possible shared computations of properties
between shadows of elements of the Experiment Data API and their
originals. The goal is to reduce runtime and memory consumption by
avoiding the computation of attributes which would have identical values
on shadows and originals where possible. There is still more potential,
but at least I think the amount of shared computations should now be
higher.