New Attribute: OnlySharedInstanceRuns
When comparing optimization algorithms, it is important to compare the
data fairly. This means that if we compare algorithms A and B, where
we have data for benchmark instances 1, 2, and 3 for A while having
data for benchmark instances 1 and 3 for B, we can only compare A and
B based on instances 1 and 3. This sounds kind of natural, but if we
compare, for instance,  Estimated Cumulative Distribution Functions
(ECDFs), we will usually compute them separately for each algorithm.
Thus, we may fall into the trap of computing the one of A over
instances 1, 2, and 3 - as it would be entirely correct if we just
wanted to see the ECDF for A alone. However, for comparison purposes, we
need to limit this computation to instances 1 and 3.

The new attribute builds a shadow experiment set which only contains
instance runs that belong to instances for which data exists in all
experiments of the experiment set.