Matrix Iterator Completed & Documented
We can now iterate over a set of matrices along one dimension and obtain
all values corresponding to a value of that dimension in a row matrix.
The first dimension must be rising, no restrictions are imposed on the
second dimension. The implementation tries to work with maximum
accuracy, i.e., detects whether to use long or double to represent the
data.

This approach can later be used to build compact and complex functions
over data from the Experiment Data API. Namely, I will use it to join
several ECDFs from different IInstanceRuns.