BugFix for _ShadowElementSet and Improved ECDF
The _ShadowElementSet is supposed to automatically set the owners of
un-owned shadow elements in its set, but due to a wrong "instanceof",
it would do so only if the elements were _ShadowProperty-s. This is
fixed now, and the potential NullPointException on getOwner().something
does not happen anymore.

The ECDF calculation is improved: It now gives all IInstanceRunSets and
IExperiments the same weight during aggregation. Imagine that you have
20 runs for the first benchmark instance and 40 for the second instance.
At a given time step, 10 runs for experiment 1 are successful and 10 for
experiment 2. The ECDF at this time step should be 0.5*(10/20 + 10/40) =
0.5*(0.5 + 0.25) = 0.5*0.75 = 0.375. In the previous version, it would
have been (10+10)/(20+40) = 20/60 = 0.3. In this old version, the second
IInstanceRuns would have received a higher weight, simply because there
are more runs for it. However, if we assume that the aggregation somehow
represents that we do not know which kind of problem we would solve in
practice beforehand, i.e., that each problem could occur with the same,
uniform probability, then each instance should receive the same weight.
This is implemented now properly (by using the new MatrixIterator2D
committed earlier today).