FunctionJob Only Uses Instances for which Data Exists in All Experiments
The FunctionJob base class to develop experiment set modules for
plotting functions over several experiments for comparison purposes now
uses only data from instances for which results exist in all
experiments. This is the fair approach and this data is obtained via the
new attribute OnlySharedInstanceRuns. The data is now pre-computed in
its entirety, allowing us to know how many diagrams we will plot in
advance.

The implementation of the Estimated Cumulative Distribution Function
(ECDF) experiment set module is the first (and currently only) module
based on this approach. It seems to work. But will need to be tested
more thoroughly.