Better Division + Some USeless Code Deleted
The division operation has been improved: If we divide two values which
can be represented as long and the result is a fraction, we do not
divide the values directly. Instead, we first compute their greatest
common divisor (GCD) and divide both with that. Then result is computed.
It will, obviously, still be the same fraction. However, since the
actual division takes place in (double) floating point, it seems logical
that a division of smaller values may lead to better precision results.
At least it will not be worse. Or at least should not be worse.

Some useless code has been deleted in the statistics package.