Improvements to NumericalTypes, Drafts for Mean and Variance Aggregates
The NumericalType class now sports a method for directly checking
whether doubles or floats can be represented as longs. This should save
some time in cases where the full getTypes is not needed.

As a result, several mathematical functions have been modified.

Draft aggregates for computing the Arithmetic Mean and the Variance on
the fly have been added. The focus is on accuracy and I don't care
whether I have to sacrifice runtime for this. This means that we use
StableSums where ever possible as well as long arithmetic whenever we
can. This, however, results in much larger code, as we have to deal
with states such as overflow oder infinity etc. For the Arithmetic Mean,
I implemented that more or less, for the Variance aggregate, I didn't do
it yet.

I decided to delete the class StatisticInfo and its required internal
classes. I think it is better to have dedicated aggregates for different
statistical parameters. At least for now.