Aggregate for Quantiles
An aggregate for quantiles has been added. It caches all values in an
internal array and computes actual quantile values when the state is
requested via getState, doubleValue, or longValue. The validity of the
aggregate is checked with a JUnit test comparing it the apache commons
Percentile class. Different from this class, we support long values
data and try to use our (hopefully) more precise arithmetic operations.