StatisticalParameter: Parsable ISemanticComponent for Statistical Params
The new class StatisticalParameter has been defined as base class for
specifying, well statistical parameters. It provides the methods of
ISemanticComponent to append its name to text output devices. It also
can create a ScalarAggregate for computing its value over a data set.
Instances of it can furthermore be parsed with the new
StatisticaParameterParser. This will allow us to configure which
statistical parameter should be computer over a data set in the
experiment data API. This will, in turn, allow us specify a wide
variety of diagrams to be painted based on this.