First Version of Aggregate 2D Function + BugFix for Mul
The evaluator module for plotting diagrams of how an aggregate
such as the median of a solution quality changes over another
dimension such as time has been created. The data to be plotted
can be chosen virtually arbitrary.

A bug fix in the mathematical function Mul for multiplication has
been introduced: The original code could feature an aweful
division by zero if you were to multiply with 0. Argh.