First Approach to Document API Macros
A first approach to macros for the Document API has been created. A
macro here is a recording of methods applied to an instance of the
Document API which can be re-played (applied) to other instances of
the same interface.

The idea is as follows: All elements of the Document API are specified
as interfaces and extend Closeable. If we want to record a macro, we
create a dynamic Proxy of the interface we desire to record from. The
InvocationHandler of the proxy then records all method calls and (deep
clones of) their parameters in a list. If a method is called which is
supposed to create a new instance of a Document API interface, we can
dynamically create a proxy for that instance as well. When the original,
root proxy object is closed (via its close() method), we can then take
the list of method calls and apply it to arbitrary instances of the that
interface. Notice that the final call to the close() method is not
recorded, so that the other instances we play the macro to are not
automatically closed by it.

This is a relatively straightforward approach, which should work well,
but also has some issues:
1. We may clone char[] arrays which are unnecessarily long if they are
   parameters to a method supposed to only print a sub-sequence of them.
2. During the macro recording, no errors will be generated. Exceptions
   will be thrown once the macro is re-played.