BugFix Graphics: Draw and Fill Arcs, Text Scaling
The handling of arcs has been improved in Graphics (and thus also all
classes subclassing it). In the old version, arcs were drawn wrong in
the new _PGFGraphic and twice in all other graphics.

The scaling of text in the _PGFGraphic has been improved. Text now looks
pretty like what we would expect based on Java metrics, even though it
is rendered and kerned by LaTeX.