First Steps towards an "Export" Document Type
When evaluating experimental data, we maybe do not just want to create
documents, such as PDFs or web pages, with the evaluation results.
Instead, maybe we want to export some of the evaluation results to
other applications, such as GnuPlot or Matlab. Such applications are
usually not suited to load text data or charts from PDFs or XHTML files.

The Document API is, at first glance, not suitable to produce textual
output which is easy to parse. However, with some hacks, I can get
this to work:

1. I introduce a new graphic driver which actually does not render
   graphics but instead implements ITextOutput, the abstract interface
   to write text. It stores all text to a text file. This text file
   is then its final "product" (instead of a graphic).
   
2. I add a new chart driver which, instead of rendering a diagram to a
   graphic, stores the chart data in a text format similar to what
   GnuPlot reads to that ITextOutput. Of course, this chart driver
   will only work together with the text graphic driver.
   
3. A new implementation of the Document API is generated which makes
   use of these graphic and chart drivers. This way, it will store all
   data which would otherwise be rendered as a chart as text in one
   text file per chart.
   
This document driver will, in the future, be extended to also produce
one text file per table. We can then render tables to text files and
load them into other applications. Furthermore, it should also render
each source code listing to a separate text file.

Inside the main text file, it should merely list the section titles
and, for each table, figure, or listing the title and the path to
the corresponding file. If we keep the format like this, it could
even be parsed by another application entirely.

While points 1 to 3 above have been realized, the ideas in the two
paragraphs following them have not yet been implemented. I will do so
soon(er or later ^_^).