Improved Compound Functions and Function Rendering
We finally have some good support for creating compound mathematical
functions. In our "functions" package, functions are implemented as
singleton instances. With the new compound support, we can chain and
combine arbitrary such instances in an efficient way. We also support
parameters and constant functions. Most of the code of this new approach
is automatically generated and the _CodeGenerator is also provided
along in the package, in case that something changes in future.

Additionally, with the new interface IParameterRenderer, we can also
render functions to text (ITextOutput) or the Document API (via IMath).