Parser for Compound Mathematical Functions
A parser for compound mathematical functions has been introduced. This
parser uses the builder API for compound functions and can basically use
any mathematical function either provided in our package or residing
in the classpath. It also uses the interface INameResolver to resolve
names of variables or other external objects.

The goal behind this is to enable us to have functions applied to
experimental data based on arbitrary transformations of the input and
output data.