Another Intermediate Commit
This is an intermediate commit on our way to fully user-definable
transformation functions for data. We can now parse basically arbitrary
mathematical expressions, containing numbers, constants, dimension or
parameter/feature names to compound expressions. We can also render
expressions to text more or less well.

However, in the process, the idea of ISemanticComponent needs to be
re-defined. Due to the now much more complex (but also unified)
structure of the data transformations we can get, plus the semantically
more strict distinction between mathematical and non-mathematical stuff,
I need to find a better way to have objects which can describe
themselves to text.

Since many of the core objects are "old" instances of the "old"
ISemanticComponent, several methods have to be re-designed. The changes
go into the core of our system, so on the way, some things will break.
However, I want to make intermediate commits to have some save location
I can go back to, just in case. Thus, be adviced, this commit is broken.