Another Intermediate Commit on the Way to Transformation Functions
This is (hopefully) the laste intermediate commit on the way to allowing
for arbitrary-nature tranformation functions to be applied to input
data. Now a concept for the ECDF with input and output transformations
has been developed as well.

Still, there are compiler errors and stuff, but these will be fixed
(hopefully tomorrow).