BugFix for Template Example
The template example contained the now-removed element "nprop" for "not
proportional to". Although the Unicode standard has a character for
"proportional to", it does not have one for the opposite. Dealing with
this (probably never used) sitation required ugly hacks, so I removed
it.

The Maven POM has been updated to the target version 0.8.1.