Changed Rendering of Parameters
The way parameters of a mathematical function are rendered is now a tiny
bit different: Parameters receive the highest precedence priority. Based
on the precendence priority, we decide whether or not the arguments of
a function should be put into braces. For example, we would put braces#
in (5+3)*5 because the root node of the argument "5+3" of the
multiplication, the "+", has a lower precedence priority than "*" and
without braces, the "*" would "steal" the 3. We do not put braces in
5+(3*5), i.e., render 5+3*5 directly, as the precedence is clear here.

For parameters, such as "F" in "F/5", we now assume the highest
precedence priority. This looks good so far, as otherwise you might get
"(F)/5". But it _may_ have implications later: It could be that a whole
function is rendered as parameter, in which case the function rendering
should put braces by itself. Otherwise, the case discussed above, that
a stronger outer function might "steal" inner arguments, may occur.