Create Composite Fonts to Fix Missing Glyphs
With our evaluator, we want to create algorithm performance reports with
as publication-ready-as-possible components. This means that figures
should use the fonts of the document class, when reports are rendered as
LaTeX documents.

The default fonts in LaTeX often are Knuth's Computer Modern (Roman)
fonts and their derivates. Unfortunately, these fonts have a limited set
of glyphs and do not support the full Unicode standard (or at least a
large enough subset for reports). They do, for instance, not have glyphs
for "<", the "less-or-equal"-symbol, "²", or "{" and "}". Now these are
fairly standard symbols that a user might include in a chart legend or
title, and I have little means to prevent them from doing so.
Unfortunately, such symbols will then be rendered as empty rectangles,
looking something like "[]". This kind of messes up the experience and
destroys the concept of "publication-ready".

I have three choices to deal with this:
1. Ignore it (not good).
2. Create and use versions of the CMR fonts which support the full
   Unicode range. This might be a good idea, but a) would I need to do
   that for all fonts and all document classes I support, b) I need to
   find such fonts at reasonable licensing, c) would the fonts then
   still not be the exact fonts used in the paper but copies thereof.
3. Try to change the way text is rendered in my implementation of
   Graphics. Now this would be tedious, I would need to split strings
   manually at missing gylphs, load the right fonts for the missing
   glyphs, and lay them out manually. Also, this solution only works
   for my own graphics, not for any other, user provided ones.
4. Create composite fonts using the "public" available utility classes
   of Java. Via Swing's StyleContext, I can obtain a font in form of
   a UI Resource which matches an existing font but is a composite
   providing glyphs for anything the existing font does not.
   Unfortunately, these glyphs tend to be in the Dialog font, i.e.,
   sans-serif, and might look quite different from the original font's
   other glyphs. I have no control over where the missing glyphs come
   from.
5. Use the classes in the sun.something packages, which are partly
   package-private, but via which I would gain that control.
   Unfortunately, there would be a very high chance to break the
   behavior in future Java versions.

I chose point 4. It is not really beautiful, but for now it will do. In
the future, I may chose a 4./5. hybrid, trying to do 5. but using 4. as
fallback. Or maybe the problem will solve itself and a Java version
after Java 7 provides me with the ability to do 5. in a "legal" way.