Better Textual Descriptions of Evaluation Elements
A lot of work has been done to allow for better textual descriptions in
the sections generated by evaluator modules. The ECDF and Aggregation2D
evaluation jobs add better text, mainly due to improvements of the
descriptions generated by the function attributes backing them.

Some minor changes have been done as well: The Document API now allows
for the allocation of labels from section bodies, the PathUtils'
sanitizePathComponent is improved, labels of figures and sub-figures
are stored in the data objects of the function jobs, ETextCase supports
getting the next case for after the current sentence, and TextUtils'
normalize method allows for specifying a Normalizer.Form.