Sqr, Sqrt, Cube, Cbrt now have 1-char ITextOut mathRender
We now can include arbitrary unicode characters into text strings of
arbitrary fonts (as we always use composite fonts and do not depend
on the limited character range of a given font anymore), as of the
previous commits. Therefore, we chose to represent Sqr, Sqrt, Cube,
and Cbrt by the corresponding unicode characters instead of function
names (no longer "sqrt(x)") and powers ("²" instead of "x^2").