Several Changes to Increase Robustness
Normally, when loading experiment data into the evaluator, we will
compile data container classes which perfectly fit to the measure
dimensions on the fly. This allows for fast access to the data and a
small memory footprint.

However, if no Java compiler is available, because a JRE is used to
execute the program instead of a JDK, we failed with an exception.

Now, fallback container classes are provided. These classes are both
slower and need more memory than those we could automatically generate,
but we can now also run under a JRE.

Further improvements include dealing with null file ID objects (which
may happen under Windows), better logging output for the CSV+EDI format,
and slightly improved documentation.

This will become release 0.8.2.