Overhaul of the LaTeX Tool Chain
I did some cleaning up of the code of the LaTeX Tool Chain. The
resulting classes are now somewhat more compact with more shared code.

Also, we now support a LuaLaTeX-based tool chain. This is importang
since we also support generating graphics in the PGF format, a LaTeX
dialect. Now large PDF figures may lead LaTeX, XeLaTeX, PdfLaTeX et al.
to fail with out-of-memory errors. LuaLaTeX seems to be able to
dynamically allocate more memory if necessary and hence, to be able to
compile such documents.
See http://tex.stackexchange.com/questions/7953

For this reason, the new tool chain is prefered, i.e., when the LaTeX
Tool tries to find a suitable compiler for a given constallation of
file types (tex, bib?, pdf?, eps?, png? ...), it will first try to
detect a LuaLaTeX or LuaTeX compiler. If none is found, we will look for
other alternatives.