Modification of ExperimentSetTest
A modification of the method testExperimentRunsFindValuesBetween of
ExperimentSetTest. There was an error in a Travis CI build
https://travis-ci.org/optimizationBenchmarking/optimizationBenchmarking/jobs/71506962
which may have been caused by the structure of the test which may not
have caught a special condition for very large numbers.

However, I am not sure about this issue, so I will need to keep an eye
on this.