First Approach to Ranking + Removed ETextCase.ensure
The first draft for ranking has been introduced. It is similar to the
ranking in Apache commons.math3, but different from it, it can
consistently rank both `double` and `long` data. Since there is no way
to directly compare certain values (that is, very large `long` values to
similarly large `double`s), we convert some values to BigDecimal if
needed. The used classes for the ranking are designed such that it will
also be possible to rank two-dimensional data, which will come in handy
when we implement the Mann-Whitney U test. This, in turn, will be
necessary since it requires ranking and the implementation from Apache
commons.math3 only supports `double` values, while the data we deal
with may potentially contain large `long`s (and, if I understand
correct, converting any `long` larger than 2^52 to `double` will
lead to a loss of fidelity).

Furthermore, we removed the method `ensure` from `ETextCase`. I rather
have some `NullPointerException`s and sort them out than continuing to
clutter my code with this.