FunctionJob now supports Ranking Transform and Many Other Things
1. FunctionJob now can transform its functions to rankings. Sometimes,
you may not be able to distinguish which experiment has higher and which
has lower function values. If you have three experiments in a plot, and
two have small function values while one has large values, you probably
cannot see which of the former two has the smallest value. By instead
plotting a ranking of the values, the highest value would be 3, the
smallest one 1, and no two values can be closer to each other than 0.5.
In other words, you can very clearly see when which algorithm is better
and when which algorithm takes over another one.

2. The concept of having additional start and end points in plots has
been removed. This feature was not used yet and may only need to
complications without any reward.

3. Some bugs have been fixed in MatrixIterator2D.