First Steps to Statistics at Specific Points
This is the first step to compute statistics at specific points in times
over an InstanceRuns set and to aggregate them over experiments and
benchmark instances.

BasicNumber has been extended with a method "toNumber" which will pack
its data into a java.lang.Number. This makes sense when translating a
statistical parameter gathered with an aggregate into a (small) object.