Path Resolution Problem: First Steps
At some point in time, paths in the program must be normalized and
transformed to absolute paths. This is, well, obvious. The normal
approach to this is to resolve relative paths towards the current
directory and keep absolute paths as they are.

However, this poses a problem: What if I have a configuration file
containing relative paths which should be resolved based on the
configuration file's parent folder instead of the current folder of the
application? This problem emerges, e.g., when we want to run evaluation
procedures based on configuration files in different folders. It
appears, in particular, in the GUI components I am working on.

In order to solve this problem, our APIs for input and output have to
be extended to accept a base path against which relative paths should
be resolved. Not setting the base path or setting it to null should
yield the original behavior of resolution against `.`.

With this commit, I push the first attemp to have this extended behavior
for Input and Output tools. I will have to also implement it for
document drivers and the evaluation builder.