ConfigurationBuilder Now Supports Setting Base Path
We can now set a base path to be used for path resolution in a
configuration builder. If the configuration has an owning configuration,
the owning configuration must be set first. During setting the owning
configuration, we will use its base path. The base path of the root
configuration is `null`, i.e., path parameters are resolved against
'`.`' by default. If we set the base path of a configuration (builder),
this overrules inherited base paths.

IOJobBuilders inherit the base path of the configuration they are
configured with.

This should enable us to leak a base path into all elements of an
evaluation job.