Iterable Iterators and Improved Browser Selection
The new class `IterableIterator` implements both `Iterator` and
`Iterable`. An instance of it thus can convenient be used in a 
 `for( x : new MySpecialIterator()) { ... }` 
scenario. This allows me to simplify some of the existing code.

The browser selection has been updated, since I found out that we
can also reliable wait for chromium-browser under Linux and for
Chrome under Windows.