Browser Tool Now Supports java.awt.Desktop
The browser tool can now also use the
`java.awt.Desktop#browse(java.net.URI)`
component to open a web browser, if that is possible. This is considered
to be a fallback approach, since we get no handle to the web browser
process at all, so we can neither wait for it nor close it.