Numerical Parsers now have Better `hashCode` and `equals` Methods
The numerical parsers now have better implementations of `hashCode`
and `equal`. The hash code is based on the bounds and a type-specific
prime constant. This should avoid hash clashes of parsers for different
types (e.g., `LongParser` and `IntParser`) which happen to have the
same bounds. Furthermore, both `equals` and `hashCode` distinguish
strict and loose parsers. `hashCode` does that again via the prime
constant method, `equals` is implemented slightly differently for
strict and loose parsers.

Furthermore, any parsers which are not four `double` or `long` have
`getLowerBound()` and `getUpperBound()` methods fitting to their
primitive type.