Module Lists + Definition XML now allows for Inheritance Hierarchies
An XML format for specifying a set of modules has been created. We now
can load a list of available evaluation modules. This list is linked to
their corresponding parameter definitions, which are loaded on demand.

In order to keep the parameters and definitions clean, we now support
"inheritance" of parameters: A class can inherit parameters from several
other classes. The definition loading procedure will recursively
traverse this hierarchy.

Thus, we now have all the tools at hand to create an editor for
evaluation processes. The choices presented by this editor will
automatically change as new modules are added or as configuration
parameters change.