Now `FunctionJob` prints Figure ref, `EDIOutput` can output partial data
`FunctionJob` now includes a reference to the figure for each job it
does.

`EDIOutput` has been modified to also be able to store other elements of
the Experiment Data API than just `IExperimentSet`, namely
`IDimensionSet`, `IInstanceSet`, and `IExperiment`. This is needed to
get the GUI to work, where forms must be able to generate and store such
data structures.