Flat EDI Input Tool added + `LineIterator` much more efficient
I added an input tool using the flat experiment set context API.

The class `LineIterator` is now much more efficient. It is no longer
backed by `BufferedReader` plugged on a `StringReader`. Instead, it
directly iterates over the characters in a string. It also now supports
auto-trimming of the line elements as well as setting empty strings to
null, if required.