Attempt to Speed Up Detection of External Process Tools (LaTeX, Browser)
This is an attempt to speed up the detection of external tools (the
LaTeX installation and web browser) which needed to be detected from
the PATH. (PATH here is used in the sense of the PATH system variable
pointing to the folders with executables). For many tools, I provide
better default paths to be searched first. For the browser, some
browsers (such as Internet Exploter and Edge) can be excludeded from the
search if the operating system is not Microsoft Windows.