Draft of Text-based Process API + More Null Streams
While our `ExternalProcess` API allows the starting of and deadlock-free
communication with external processes, it is limited to binary streams.
In other words, you can write bytes to the stdin and read bytes from the
stdout of these processes. Some processes, such as shells, however,
require text-based streams. I therefore generalized the process API to
also provide `TextProcess`es, with which we can communicate via
`BufferedReader`s and `BufferedWriter`s. These `TextProcesses` are
wrapped around the binary `ExternalProcess`es. The corresponding builder
takes the same arguments as the one for normal processes.

Additionally, there is a new interface called `IProcessCloser`. For each
new process, you can specify an instance of this interface, whose
`beforeClose` method is called after the user is done with the process
and either invokes `waitFor()` or `close()`. This allows us to perform
a graceful shutdown. If you have opened a shell, you could provide an
implementation of this interface, which writes `exit` followed by a
newline to the process, in order to achieve graceful termination.

Additionally, more streams which directly discard their output or
provide no input are provided, such as `NullBufferedWriter` and
`NullBufferedReader`.