Improved Process API + "Shell" Tool
The process API has again been improved.

There now is a "Shell" tool which allows the creation of a system shell
on which several commands can be executed and whose output can be
caught. It is based on the new `TextProcess` API which I will hopefully
also be able to use for a better communication with `R`.