Total Clean-Up of the MathEngine API
That API was totally over-engineered, with functionality which I
actually don't need. The goal of this API is to provide an interface to
communicate with other mathematics software, such as R. Iwant to use the
machine learning and data mining functionality of R, to be precise. For
this, all I need is to send some data from my software to an R process,
execute a script in R, then read back the result. The sending and
receiving of data has now be implemented, the script will follow soon.

The previous version of the API had functionality such as transposing
and inverting of matrices. All of that is actually not needed, as it can
be done inside the scripts. There it can also be done much more
reliable, since we can deal with errors in a script appropriately and do
not need to worry what to do with them in our system.