Added Line Iterator for Streams and Script Capability for Math Engines
The `IMathEngine` interface now allows for the execution of scripts.
Scripts are provided in the form of an `Iterable<String>`, i.e., a
sequence of string commands. In the `REngine` implementation of that
interface, these lines/commands are written directly to the stdin of the
`R` process.

I furthermore added an implementation of the interfaces
`Iterable<String>`, `Iterator<String>` (`Iterable.iterator()` returns
`this`), and `Closeable` which can draw its data from a
`BufferedReader`, `Reader`, `InputStream`, or resource. It will preserve
a potential `IOException` caught during the iteration process and
re-throw it on `close()`.

The above two building blocks together allow me to store an `R` script
as resource and to execute it by conveniently loading said resource with
the new iterator.

Based on this, a first draft of a test case for the `REngine` has been
developed, which loads a simple script only inverting a matrix (provided
in a variable) and storing the result in another variable.