`REngine` doesn't need temp files anymore and better respects `int`
The `REngine` should now be faster, as it does not need temporary files
anymore for communication. So far, large matrices were stored in temp
files when communicating with `R`. This is now not necessary anymore, as
we write data separated by newlines to the stdin.

Furthermore, by using `as.integer`, we try to better respect
integer-ness of data. Since `R` does not support 64 bit integers
natively, we only can only create 32bit integers like that.