First Step towards Clustering by Performance
The performance of an algorithm on a given benchmark instance can be
represented by a fingerprint of basic statistics computed over the
raw log data. These statistics are based on the first and last point
of the run, as well as on some points in between.

I want to use these fingerprints for clustering
  1. benchmark instances by the performance of the algorithms on them
  2. algorithms based on their performance on the benchmark instances
  
Since there may be a large amount of data, this will probably involve
the following steps:

 1. [in R] normalizing of data?
 2. [in R] feature selection
 3. [in R] clustering
 4. determining of the relationship between feature values and clusters
 
 The first three steps will (hopefully) be done in the R language via
 the newly implemented interface to R. This way, I can use readily
 availble data mining and clustering approaches.