First Steps towards Performance-based Instance Clustering
This commit includes the first steps towards providing a new way to
group benchmark problems: Performance-based Clustering of Benchmark
Instances. Here, we first compute some key statistics for all the
instance runs available for an instance. Thus, we have a row vector of
raw statistical data per instance. All of these row vectors form a
matrix which is then handed to the REngine. We use R to first perform a
principal component analysis to reduce the features. Then we try to find
a good clustering (with a large silhouette value) by testing different
algorithms and different target cluster numbers. The best discovered
clustering is then used to group the instances.