Next Step towards Clustering (of Instances) by Algorithm Performance
I now attempt to fit a model function to the way an algorithm performs
on an instance based on the log data of all runs of that algorithm on
said instance. This will hopefully be more accurate as fitting only
based on key statistics. This is done for each dimension pairing.

If we model the objective value over time, then I currently try to use
an exponential decay model of the type

 performance = -[ exp( -A/(time^(-B) )  -  1]
 
which is fitted on normalized performance and time dimension.
Normalization seems to be necessary because otherwise, the exponent
and power in the formula make it go to infinity too quickly.

If a time and another time or an objective and another objective
dimension are matched, we use a simple polynomial a+b*x+c*x^2 as model.

For function fitting, a new tool has been introduced. Currently, this
tool's implementation uses some of Apache Commons Math 3's technologies.
It starts by picking the best choice out of several random fittings.
Then it optimizes this with the Levenberg-Marquardt algorithm (if that
fails, with Gauss-Newton based on SVD). Then it applies the Nelder-Mead
simplex to the result for further improvements (if that fails, BOBYQA).
Active CMA-ES is used as afterburner.

As objective function for this, we use the sum of the worst M errors
from all the N>>M points. All errors are scaled. This is to deal with
the situations where 
  a) we have many points on one end of the axis and few at the other
     and would otherwise ignore the few (if using the sum over all
     errors), and
  b) we may have objectives with exponential behavior, i.e., which may
     start at 100, go down to 10, then to 1, to 0.1, to 0.01, and so on
     during the optimization process - and without error scaling, we
     would ignore the smaller points.