Fixed Bug in `ArraySetView`
`ArraySetView` would throw a `ClassCastException` when looking for an
object which does not fit to the Comparable interface implementations of
its elements. This would happen, e.g., when looking up for an object
which is not the right type of `Enum`s in a list of `Enum`s. It
concerned the `indexOf` and `contains` methods. It made JUnit tests for
the LaTeX Tool Chain crash.