Trying to Prevent useless Logging Entries from XToolkit
The default filter added to loggers and handlers. It seems that the
internal implementation of the graphics toolkit for X11, namely
"sun.awt.X11.XToolkit", generates a lot of logging entries of level
FINER talking about some timeout tasks being or being not executed.
This happens at such a frequency that it just fills up the logs with
rubbish. With this filter here, we try to prevent this behavior. The
filter is applied to the global logger and all of its parents and
its/their handlers recursively. It will hopefully remove this useless
log entries. The filter is applied in the moment when we invoke
org.optimizationBenchmarking.utils.config.Configuration#getGlobalLogge
r().